chore(admin): convert layouts to TS
