+++ packages/core/admin/admin/src/pages/Admin/index.js
@@ -16,7 +16,7 @@ import { GuidedTourModal } from '../../components/GuidedTour/Modal';
 import { LeftMenu } from '../../components/LeftMenu';
 import { useConfiguration } from '../../hooks/useConfiguration';
 import { useMenu } from '../../hooks/useMenu';
-import AppLayout from '../../layouts/AppLayout';
+import { AppLayout } from '../../layouts/AppLayout';
 import { createRoute } from '../../utils';
 import { SET_APP_RUNTIME_STATUS } from '../App/constants';
