+++ packages/core/admin/admin/src/pages/AuthPage/components/ForgotPasswordSuccess/index.js
@@ -5,9 +5,10 @@ import { Link } from '@strapi/helper-plugin';
 import { useIntl } from 'react-intl';
 
 import { Logo } from '../../../../components/UnauthenticatedLogo';
-import UnauthenticatedLayout, {
+import {
   Column,
   LayoutContent,
+  UnauthenticatedLayout
 } from '../../../../layouts/UnauthenticatedLayout';
 
 const ForgotPasswordSuccess = () => {
