+++ packages/core/admin/admin/src/pages/AuthPage/components/Login/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import PropTypes from 'prop-types';
 
-import UnauthenticatedLayout from '../../../../layouts/UnauthenticatedLayout';
+import { UnauthenticatedLayout } from '../../../../layouts/UnauthenticatedLayout';
 
 import BaseLogin from './BaseLogin';
